requirements.txt
aiogram==3.12.0
python-dotenv==1.0.0
httpx==0.27.0
aiofiles==23.2.1
urllib3==2.1.0
requests==2.31.0
charset-normalizer==3.3.2
idna==3.6
